refactored common and validator structure, now only use index.ts for export marking
