+++ packages/common/index.ts
@@ -1,2 +1,1 @@
-export * from './guards';
-export * from './types';
+export * from './src';
