+++ packages/common/src/guards/isRecord.ts
