+++ packages/common/src/index.ts
@@ -1,0 +1,3 @@
+export * from './guards/isRecord';
+export * from './types/flatten.types';
+export * from './types/prettify.types';
