+++ packages/common/src/types/flatten.types.ts
