+++ packages/common/tsconfig.json
@@ -1,10 +1,10 @@
 {
 	"extends": "../../tsconfig.base.json",
     "compilerOptions": {
-        "outDir": "dist"
+        "outDir": "lib"
     },
     "exclude": [
         "node_modules",
-        "dist"
+        "lib"
     ]
 }
