--- packages/common/types/index.ts
@@ -1,2 +1,0 @@
-export * from './flatten.types';
-export * from './prettify.types';
