+++ packages/validator/.gitignore
@@ -130,3 +130,5 @@ dist
 .yarn/build-state.yml
 .yarn/install-state.gz
 .pnp.*
+
+lib
