+++ packages/validator/src/typebox/typeboxSchemaValidator.ts
@@ -29,7 +29,7 @@ import {
   LiteralSchema,
   ParseResult,
   SchemaValidator as SV
-} from '../types/schema.types';
+} from '../shared/types/schema.types';
 import {
   TCatchall,
   TIdiomaticSchema,
