+++ packages/validator/tests/zod/schemaValidator.test.ts
@@ -1,184 +1,22 @@
 import { generateSchema } from '@anatine/zod-openapi';
 import { ZodObject, z } from 'zod';
-import { Schema } from '../../index';
-import { UnboxedObjectSchema } from '../../types/schema.types';
+import { UnboxedObjectSchema } from '../../src/shared/types/schema.types';
 import {
   array,
-  bigint,
-  boolean,
-  date,
-  empty,
-  never,
   number,
   openapi,
   optional,
   parse,
   schemify,
   string,
-  symbol,
   union,
   validate
-} from '../../zod';
-import { ZodObjectShape } from '../../zod/types/schema.types';
-import { ZodSchemaValidator } from '../../zod/zodSchemaValidator';
+} from '../../src/zod';
+import { ZodObjectShape } from '../../src/zod/types/schema.types';
+import { ZodSchemaValidator } from '../../src/zod/zodSchemaValidator';
 import { compare } from '../utils/compare';
 
-const one = array({
-  name: {
-    j: union([string, number, date, boolean, bigint, empty, symbol, never]),
-    t: optional(
-      union([
-        array({
-          y: array(number)
-        }),
-        string
-      ])
-    ),
-    m: {
-      a: optional(string),
-      b: number,
-      c: {
-        d: string,
-        e: number
-      }
-    }
-  },
-  200: {
-    j: string
-  },
-  m: {
-    a: true as const
-  }
-});
-const two = array({
-  name: {
-    j: union([string, number, date, boolean, bigint, empty, symbol, never]),
-    t: optional(
-      union([
-        array({
-          y: array(number)
-        }),
-        string
-      ])
-    ),
-    m: schemify({
-      a: optional(string),
-      b: number,
-      c: {
-        d: string,
-        e: number
-      }
-    })
-  },
-  200: {
-    j: string
-  },
-  m: {
-    a: true as const
-  }
-});
-const three = schemify(
-  array(
-    schemify({
-      name: schemify({
-        j: union([string, number, date, boolean, bigint, empty, symbol, never]),
-        t: optional(
-          union([
-            array({
-              y: array(number)
-            }),
-            string
-          ])
-        ),
-        m: schemify({
-          a: optional(string),
-          b: number,
-          c: {
-            d: string,
-            e: number
-          }
-        })
-      }),
-      200: schemify({
-        j: string
-      }),
-      m: schemify({
-        a: true as const
-      })
-    })
-  )
-);
-
-export function assert<T extends never>() {}
-type Equality<A, B> = Exclude<A, B> | Exclude<B, A>;
-
-type SchemaOne = Schema<typeof one, ZodSchemaValidator>;
-type SchemaTwo = Schema<typeof two, ZodSchemaValidator>;
-type SchemaThree = Schema<typeof three, ZodSchemaValidator>;
-
-type Expected = {
-  name: {
-    j?:
-      | string
-      | number
-      | bigint
-      | boolean
-      | symbol
-      | void
-      | Date
-      | null
-      | undefined;
-    t?:
-      | string
-      | {
-          y: number[];
-        }[]
-      | undefined;
-    m: {
-      a?: string | undefined;
-      b: number;
-      c: {
-        d: string;
-        e: number;
-      };
-    };
-  };
-  200: {
-    j: string;
-  };
-  m: {
-    a: true;
-  };
-}[];
-
-assert<Equality<SchemaOne, Expected>>();
-assert<Equality<SchemaOne, SchemaTwo>>();
-assert<Equality<SchemaOne, SchemaThree>>();
-assert<Equality<SchemaTwo, SchemaThree>>();
-
-const shortOne = {
-  s: string,
-  non: number
-};
-
-const shortTwo = schemify({
-  s: string,
-  non: number
-});
-
-type ShortExpected = {
-  s: string;
-  non: number;
-};
-assert<Equality<Schema<typeof shortOne, ZodSchemaValidator>, ShortExpected>>();
-assert<
-  Equality<
-    Schema<typeof shortOne, ZodSchemaValidator>,
-    Schema<typeof shortTwo, ZodSchemaValidator>
-  >
->();
-
-describe('Zod Equality Tests', () => {
+describe('zod schema validator tests', () => {
   let schema: UnboxedObjectSchema<ZodSchemaValidator>;
   let schemified: ZodObject<ZodObjectShape>;
   let expectedSchema: ZodObject<ZodObjectShape>;
@@ -203,7 +41,7 @@ describe('Zod Equality Tests', () => {
     });
   });
 
-  test('Schema Equality', async () => {
+  test('schemify', async () => {
     expect(compare(schemified, expectedSchema)).toBe(true);
 
     expect(
@@ -260,7 +98,7 @@ describe('Zod Equality Tests', () => {
     ).toBe(true);
   });
 
-  test('Optional Schema Equality', async () => {
+  test('optional', async () => {
     const unboxSchemified = optional(schema);
     const boxSchemified = optional(schemified);
 
@@ -269,7 +107,7 @@ describe('Zod Equality Tests', () => {
     expect(compare(boxSchemified, schemifiedExpected)).toBe(true);
   });
 
-  test('Array Schema Equality', async () => {
+  test('array', async () => {
     const unboxSchemified = array(schema);
     const boxSchemified = array(schemified);
 
@@ -278,7 +116,7 @@ describe('Zod Equality Tests', () => {
     expect(compare(boxSchemified, schemifiedExpected)).toBe(true);
   });
 
-  test('Union Schema Equality', async () => {
+  test('union', async () => {
     const unboxSchemified = union([
       schema,
       {
@@ -317,7 +155,7 @@ describe('Zod Equality Tests', () => {
     expect(compare(boxSchemified2, schemifiedExpected)).toBe(true);
   });
 
-  test('Literal Schema Equality', async () => {
+  test('literal', async () => {
     const schemified = schemify({
       hello: 'world'
     });
@@ -331,7 +169,7 @@ describe('Zod Equality Tests', () => {
     ).toBe(true);
   });
 
-  test('Validate Schema', async () => {
+  test('validate', async () => {
     expect(
       validate(schemified, {
         hello: {
@@ -353,7 +191,7 @@ describe('Zod Equality Tests', () => {
     ).toBe(false);
   });
 
-  test('Parse Schema', () => {
+  test('parse', () => {
     const validParse = parse(schemified, {
       hello: {
         world: 'world'
@@ -394,7 +232,7 @@ describe('Zod Equality Tests', () => {
     }
   });
 
-  test('OpenAPI Conversion', async () => {
+  test('openapi', async () => {
     const schemified = schemify(schema);
     const openApi = openapi(schemified);
     expect(openApi).toEqual(generateSchema(schemified));
