--- packages/validator/types/index.ts
@@ -1,1 +1,0 @@
-export * from './schema.types';
